Added TypeNum to maps, started generalized print func
maps already get offsets to the compare and print funcs associated with it's keys; now they also have an explicit indication of the type of map they are.

With explicit types associated with strings, lists & maps it should be possible to write a general print function to use to display them.